Add thread on bot startup and start the creation of the bdgro command

+++ main.py
@@ -1,4 +1,6 @@
 import json
+import os
+
 import utils
 import sqlite3
 import discord
@@ -16,6 +18,9 @@ connexion = sqlite3.connect(config["db_name"])
 cursor = connexion.cursor()
 
 channel = None
+messages = []
+
+status = False
 
 
 def tl_log(key):
@@ -26,8 +31,14 @@ def tl_msg(key):
     return tradlib.get_translation(language, ["msg", 0, key])
 
 
-def log_command(user, command):
-    print(tl_log("command").format(user, command, utils.get_date_time()))
+def tl_thread(key):
+    if key == "":
+        return tradlib.get_translation(language, ["thread", 0])
+    return tradlib.get_translation(language, ["thread", 0, key])
+
+
+def log_command(user, command, *args):
+    print(tl_log("command").format(user, command, args, utils.get_date_time()))
 
 
 @bd_bot.event
@@ -38,22 +49,60 @@ async def on_ready():
 
     print(tl_log("on").format(bd_bot.user.name, bd_bot.user.discriminator))
 
+    channel = bd_bot.get_channel(int(config["channel_id"]))
+
+    if not channel or not isinstance(channel, discord.TextChannel):
+        print(tl_log("channel_error").format(config["channel_id"]))
+        print(tl_log("off"))
+        await bd_bot.close()
+        return
+
     utils.mpd(cursor)
     print(tl_log("db_on"))
-
-    channel = bd_bot.get_channel(config["channel_id"])
 
     await channel.purge(limit=None)
     for thread in channel.threads:
         await thread.delete()
     print(tl_log("purge"))
 
+    for key in tl_thread(""):
+        await channel.create_thread(name=tl_thread(key))
+        print(tl_log("thread").format(tl_thread(key)))
+
+    print(tl_log("threads").format(len(channel.threads)))
+
 
 @bd_bot.command()
 async def ping(ctx):
     ctx.message.delete()
     log_command(ctx.author, ctx.command.name)
+    if not ctx.author.guild_permissions.administrator:
+        await ctx.send(tl_msg("no_perm").format(ctx.author.mention))
+        return
+
     await ctx.send(tl_msg("ping").format(round(bd_bot.latency * 1000)))
+
+
+@bd_bot.command()
+async def bdgro(ctx):
+    global status
+
+    ctx.message.delete()
+    log_command(ctx.author, ctx.command.name)
+    if not ctx.author.guild_permissions.administrator:
+        await ctx.send(tl_msg("no_perm").format(ctx.author.mention))
+        return
+
+    if not status:
+        await ctx.send(tl_msg("open").format(
+            discord.utils.get(bd_bot.get_guild(config["server_id"]).roles, name=config["role"])
+        ))
+        await channel.edit(name=tl_msg("status").format(tl_msg("status_open")))
+        status = True
+    else:
+        await ctx.send(tl_msg("close"))
+        await channel.edit(name=tl_msg("status").format(tl_msg("status_close")))
+        status = False
 
 
 @bd_bot.command()
@@ -66,6 +115,7 @@ async def stop(ctx):
     if not ctx.author.guild_permissions.administrator:
         await ctx.send(tl_msg("no_perm").format(ctx.author.mention))
         return
+
     print(tl_log("off"))
     await bd_bot.close()
 
@@ -75,6 +125,7 @@ async def stop(ctx):
 
 
 if __name__ == "__main__":
+    tradlib.set_translations_files_path(os.getcwd() + "\\lang")
     tradlib.set_translation_files_extension(".json")
     tradlib.load_translations_files()
